rehabflow_ai/pyproject.toml
[project]
name = "rehabflow-ai"
version = "0.1.0"
description = "AI-Powered Physiotherapy Management System using Google MedGemma"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "gradio>=4.0.0",
    "torch>=2.0.0",
    "transformers>=4.35.0",
    "accelerate>=0.25.0",
    "safetensors>=0.4.0",
    "pillow>=10.0.0",
    "opencv-python>=4.8.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "reportlab>=4.0.0",
    "pyttsx3>=2.90",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.5.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]


[tool.ruff]
line-length = 100
target-version = "py311"

[tool.black]
line-length = 100
target-version = ["py311"]
